% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anhang.R
\name{anhang}
\alias{anhang}
\title{open the Appendix of Rclick}
\usage{
anhang()
}
\value{
None, opens pdf in default viewer using \code{\link{system2}}
}
\description{
Open the Appendix of my R handbook found online at
\url{https://github.com/brry/rclick}
}
\examples{
# anhang() # excluded from cran check because of external browser opening policy

}
\seealso{
\code{\link{funSource}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jul 2016
}
