% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmap.R
\name{bmap}
\alias{bmap}
\title{title}
\usage{
bmap(x = 13.12, y = 52.37, zm = 14, ...)
}
\arguments{
\item{\dots}{Further arguments passed to \code{\link{plot}}}

\item{a}{Numerical vector.}

\item{plot}{Logical. Should values be plotted? This can be turned off if
only the computation results are needed. DEFAULT: TRUE}

\item{dummy}{currently_Unused}
}
\value{
ReturnValue
}
\description{
description
}
\examples{
if(requireNamespace("leaflet", quietly=TRUE) && 
   requireNamespace("leaflet.extras", quietly=TRUE))
bmap()

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Dec 2022
}
