% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpairs.R
\name{bpairs}
\alias{bpairs}
\title{customized pairs plot}
\usage{
bpairs(df, main = NULL, pch = 16, col = addAlpha("blue"), ...)
}
\arguments{
\item{df}{Data.frame. Can contain NAs. Character columns are excluded.}

\item{main}{Title for the overall graph. DEFAULT: NULL (from input)}

\item{pch}{Point character. DEFAULT: 16}

\item{col}{Color. DEFAULT: \code{\link{addAlpha}("blue")}}

\item{\dots}{Further arguments passed to \code{\link{pairs}}}
}
\value{
invisible NULL
}
\description{
pairs plot with cor in the lower panel (can handle NAs), 
nice hist on the diagonal, nice scatterplot in the upper panel.
Based on the examples in pairs.
}
\examples{
bpairs(mtcars[,1:5])
mtcarsNAs <- mtcars[,1:5]
mtcarsNAs[2,3] <- NA
mtcarsNAs[2:8,4] <- NA
bpairs(mtcarsNAs)
bpairs(iris)

}
\seealso{
\code{graphics::\link[graphics]{pairs}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Feb 2025
}
\keyword{aplot}
