\name{cls}
\alias{cls}
\title{clear console by function - does not work!}
\description{clear console (CTRL + L) using a function call. Does not work, as \code{rcom} is not available!}
\usage{cls()}
\references{\url{http://r.789695.n4.nabble.com/how-to-clear-screen-in-R-console-td793936.html}}
\author{Berry Boessenkool, \email{berry-b@gmx.de}, Jan 2014}
\examples{
#cls()
}
\keyword{ programming }
\keyword{ utilities }
