% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPres.R
\name{createPres}
\alias{createPres}
\title{Create .Rnw presentation template}
\usage{
createPres(presname = "pres", dir = "presentation", path = ".",
  open = TRUE)
}
\arguments{
\item{presname}{Name of .Rnw file to be created. DEFAULT: "pres"}

\item{dir}{Name of directory that will contain .Rnw file and
fig_extern folder. "_1" will be appended if already existing,
see \code{\link{newFilename}}. DEFAULT: "presentation"}

\item{path}{Location of \code{dir}. Passed to \code{\link{setwd}}.
DEFAULT: "."}

\item{open}{Logical: run \code{\link{openFile}}? DEFAULT: TRUE}
}
\description{
Create folder with .Rnw presentation template and fig_extern folder.
}
\examples{
\dontrun{
createPres("Berry_Conference")
}

}
\seealso{
\code{\link{createFun}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Mar 2017
}
\keyword{file}
