% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataStr.R
\name{dataStr}
\alias{dataStr}
\title{str of datasets}
\usage{
dataStr(package = NULL, ...)
}
\arguments{
\item{package}{Package name. DEFAULT: NULL}

\item{\dots}{other arguments passed to \code{\link{data}}}
}
\value{
NULL. prints via \code{\link{message}} in a for loop.
}
\description{
Print the \code{\link{str}} of each dataset returned by \code{\link{data}}
}
\examples{

dataStr() # all loaded packages on search path (package=NULL)
dataStr("datasets") # only datasets in base R
dataStr("colorspace") # works with an installed but unloaded package

}
\seealso{
\code{\link{str}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, November 2015, in search of good datasets for teaching
}
\keyword{documentation}
\keyword{print}
