\name{deprecatedFunctions}
\alias{deprecatedFunctions}
\alias{changeAttribute}
\alias{showAttribute}
\alias{shapeZoom}
\alias{extremeStat}
\alias{extremeStatLmom}
\alias{compFiles}
\alias{combineTextfiles}
\title{functions removed from berryFunctions}
\description{These functions have been removed from berryFunctions recently. Using them will return a warning with directions to the replacing function.}
\usage{
changeAttribute(...)
showAttribute(...)
shapeZoom(...)
extremeStat(...)
extremeStatLmom(...)
compFiles(...)
combineTextfiles(...)}
\arguments{\item{...}{Arguments originally in the function}}
\value{warning}
\author{Berry Boessenkool, \email{berry-b@gmx.de}, Sept 2014}
\examples{
shapeZoom()
}
\keyword{documentation}
\keyword{deprecated}
