% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editRStudioProjects.R
\name{editRStudioProjects}
\alias{editRStudioProjects}
\title{edit recently opened Rprojects}
\usage{
editRStudioProjects(path = NULL)
}
\arguments{
\item{path}{currently ignored}
}
\value{
the path of the two modified files (invisibly)
}
\description{
edit the list of recently opened Rprojects in Rstudio
}
\examples{
editRStudioProjects()

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Sep 2025
}
\keyword{file}
