% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColumn.R
\name{getColumn}
\alias{getColumn}
\title{get column from data.frame}
\usage{
getColumn(x, df, trace = TRUE)
}
\arguments{
\item{x}{Column name to be subsetted. The safest is to use character strings
or \code{\link{substitute}(input)}.
If there is an object "x" in a function environment, 
its value will be used as name! (see upper2 example)}

\item{df}{dataframe object}

\item{trace}{Logical: Add function call stack to the message? DEFAULT: TRUE
WARNING: in do.call settings with large objects,
tracing may take a lot of computing time.}
}
\value{
Vector with values in the specified column
}
\description{
Extract columns if they are given in a data frame. 
Watch out not to define objects with the same name as x if you are using
getColumn in a function!
}
\examples{
getColumn(Air.Flow, stackloss)
is.error(getColumn(Acid, stackloss))

upper <- function(x) getColumn(x, stackloss)
upper(Water.Temp)
# upper(Water) # error (design choice: partial matching not supported)

upper2 <- function(xx) {xx <- 17; getColumn(xx, stackloss)}
stopifnot(is.error(      upper2(Water.Temp)       )) # breaks

upper3 <- function(xx, dd) getColumn(substitute(xx), dd)
upper3(Air.Flow, stackloss) # may be safer in many unusual scoping situations

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Sep 2016
}
\seealso{
\code{\link{subset}}
}

