% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instGit.R
\name{instGit}
\alias{instGit}
\title{install github package}
\usage{
instGit(pk, cleanup = TRUE, ...)
}
\arguments{
\item{pk}{Character string in the from of "user/package"}

\item{cleanup}{Remove downloaded zipfile and folder with source code. DEFAULT: TRUE}

\item{\dots}{Further arguments passed to \code{\link{install.packages}}, untested so far}
}
\description{
Quickly install a package from github without having to install devtools with all its dependencies.
}
\details{
Works only for pure R package structure repositories from the master branch.
Installs package dependencies listed in 'Imports' and 'Depends', but ignores version requirements!
Tested only on windows 7 with R3.2.2. 
Note: \code{devtools::install_github} is much more extensive! \cr
Note: drat is also much better than this quick hack.
\url{http://dirk.eddelbuettel.com/code/drat.html},
\url{https://github.com/eddelbuettel/drat},
\url{http://eddelbuettel.github.io/drat/DratForPackageAuthors.html}
Give your github users this code:\cr
\code{source("https://raw.githubusercontent.com/brry/berryFunctions/master/R/instGit.R")}\cr
\code{instGit("brry/extremeStat")}\cr
\code{library(extremeStat)}\cr
}
\examples{

if(FALSE){
instGit("talgalili/installr")
instGit("talgalili/installr", FALSE)
instGit("hadley/readxl")
instGit("mages/googleVis") # many dependencies!
instGit("twitter/AnomalyDetection")
instGit("yihui/knitr")
instGit("ramnathv/slidify")
instGit("jrnold/ggthemes")
}

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Dec 2015 + Mar/Apr 2016
}
\seealso{
\code{\link{funSource}}, \code{install_github} in each of the packages \code{devtools, ghit, remotes}
}
\keyword{package}

