% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library2.R
\name{library2}
\alias{library2}
\alias{require2}
\title{install.package and require}
\usage{
library2(name, libargs = NULL, ...)
}
\arguments{
\item{name}{Name of the package(s). Can be qouted, must not.}

\item{libargs}{List of arguments passed to \code{\link{library}} like \code{lib.loc}, \code{quietly} etc. DEFAULT: NULL}

\item{\dots}{Arguments passed to \code{\link{install.packages}} like \code{lib}, \code{repos} etc.}
}
\value{
\code{\link{message}s} help instruction.
}
\description{
install and load a package. If a package is not available, it is installed before being loaded
}
\note{
Passing a vector with packages will work, but give some warnings.
}
\examples{

\dontrun{
## Excluded fom CRAN checks. Package installation on server is unnecessary.
require2(ada)
library2("statmod")
}

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, 2014
}
\seealso{
\code{\link{install.packages}}, \code{\link{library}}
}
\keyword{package}

