% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locArrow.R
\name{locArrow}
\alias{locArrow}
\title{arrow at locator point in graph}
\usage{
locArrow(digits = 2, length = 0.1, code = 2, ...)
}
\arguments{
\item{digits}{Number of digits coordinates are rounded to with \code{\link{signif}}}

\item{length}{Length of the edges of the arrow head (in inches). DEFAULT: 0.1}

\item{code}{Direction of arrow head. DEFAULT: 2 (from first to last point clicked)}

\item{\dots}{Further arguments passed to \code{\link{arrows}}  like lwd, col etc}
}
\value{
Character string with code
}
\description{
Draw arrow at positions in a graph located by clicking and return the code to recreate it
}
\details{
Not tested across platforms yet...
}
\examples{

plot(cumsum(rnorm(60)), type="l")
## locArrow() # only do this manually in interactive() mode
## locArrow(col="blue", lwd=3)

}
\seealso{
\code{\link{locLine}}, \code{\link{locator}}, \code{\link{abline}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jun 2016
}
\keyword{aplot}
\keyword{iplot}
