% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locLine.R
\name{locLine}
\alias{locLine}
\title{abline at locator point in graph}
\usage{
locLine(h = TRUE, v = TRUE, n = 1, ...)
}
\arguments{
\item{h}{Draw horizontal line at clicked location? DEFAULT: TRUE}

\item{v}{Draw vertical line at clicked location? DEFAULT: TRUE}

\item{n}{Number of points to be clicked. DEFAULT: 1}

\item{\dots}{Further arguments passed to \code{\link{abline}} like lty, lwd, col, etc}
}
\value{
\code{\link{locator}} result
}
\description{
Draw vertical and/or horizontal lines at positions in a graph located by clicking
}
\details{
Not tested across platforms yet...
}
\examples{

plot(cumsum(rnorm(60)), type="l")
## locLine() # only do this manually in interactive() mode
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Mar 2016
}
\seealso{
\code{\link{locator}}, \code{\link{abline}}
}
\keyword{aplot}
\keyword{iplot}

