% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na9.R
\name{na9}
\alias{na9}
\title{Prepend spaces before na.strings}
\usage{
na9(nspace = 5, base = c("-9999", "-999", "-99", "-9.99", "-9.9", "-9,99",
  "-9,9"), more = "", ...)
}
\arguments{
\item{nspace}{number of spaces prepended. DEFAULT: 5}

\item{base}{basic na.string structures. Might be expanded in the future. Thankful for suggestions. DEFAULT: c("-9999","-999", "-99", "-9.99", "-9.9", "-9,99", "-9,9")}

\item{more}{allows more structures added to base. DEFAULT: ""}

\item{\dots}{Arguments passed to nothing currently}
}
\value{
Character strings
}
\description{
Returns a number of useful character strings with varying amount of spaces prepended. 
It can be used as \code{na.strings=na9()} in \code{\link{read.table}}.
}
\examples{

na9()

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jan 2016
}
\seealso{
\code{\link{paste}}
}
\keyword{IO}
\keyword{file}

