% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagePath.R
\name{packagePath}
\alias{packagePath}
\title{Base path of package}
\usage{
packagePath(path = ".", file = NULL, warnonly = FALSE)
}
\arguments{
\item{path}{Path to (or below) package directory. DEFAULT: "."}

\item{file}{Optional file name to be added to path. DEFAULT: NA}

\item{warnonly}{Logical: if no part of the path is a package, give a warning
and return the original input instead of stopping with an error.
DEFAULT: FALSE}
}
\value{
Path character string
}
\description{
Base path of package (with DESCRIPTION file), per default at current getwd.
             Derived from devtools::package_file
}
\examples{
# packagePath() # may fail on cran checks

}
\seealso{
\code{\link{getwd}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Sep 2017
}
\keyword{file}
