% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelCode.R
\name{parallelCode}
\alias{parallelCode}
\title{code chunk for parallelization}
\usage{
parallelCode()
}
\description{
message a code chunk template for parallelization with progress bar on windows.
On Linux, just use \code{pblapply(X, cl=8, FUN=fun)}
}
\examples{
parallelCode()

}
\seealso{
\code{\link{par_sapply}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Aug 2017
}
