% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfpng.R
\name{pdfpng}
\alias{pdfpng}
\title{Create pdf and png graph}
\usage{
pdfpng(expr, file, pdf = TRUE, png = TRUE, overwrite = FALSE, width = 7,
  height = 5, units = "in", res = 500, seed = runif(1, -1e+09, 1e+09),
  pdfargs = NULL, pngargs = NULL, ...)
}
\arguments{
\item{expr}{Expression creating the plot, can be included in curly braces.}

\item{file}{Character: Filename without pdf/png extension. 
Unless overwrite=TRUE, files will not be overwritten, but
"_1" will be appended instead, see \code{\link{newFilename}}.}

\item{pdf}{Logical: Create pdf? DEFAULT: TRUE}

\item{png}{Logical: Create png? DEFAULT: TRUE}

\item{overwrite}{Logical: Overwrite existing \code{file}? Can be a vector
for pdf and png separately.
DEFAULT: FALSE (_n appended in filename)}

\item{width, height}{Graph dimensions. DEFAULT: 7x5 inches}

\item{units, res}{Graph quality arguments passed only to \code{\link{png}}.
DEFAULT: inches ("in"), 500 ppi}

\item{seed}{Seed passed to \code{\link{set.seed}} before each call. 
DEFAULT: runif(1,-1e9,1e9)}

\item{pdfargs}{List of arguments only passed to \code{\link{pdf}}.}

\item{pngargs}{List of arguments only passed to \code{\link{png}}.}

\item{\dots}{Further arguments passed to both \code{\link{pdf}} and \code{\link{png}}}
}
\value{
Nothing
}
\description{
Create both a \code{\link{pdf}} and a \code{\link{png}} file with a graph,
with custom size default values.
To iteratively create pdfs without closing and reopening the pdf viewer, you 
might want to use e.g. Sumatra, which comes with Rstudio. It can be found e.g. in
C:/Program Files/RStudio/bin/sumatra
}
\examples{

pdfpng({par(bg=8, las=1); plot(cumsum(rnorm(500)), type="l")}, 
       file="dummyplot", res=100)
pdfpng({par(bg=8, las=1); plot(cumsum(rnorm(500)), type="l")}, 
       file="dummyplot", overwrite=c(TRUE,FALSE))
       
unlink("dummyplot.pdf") ; unlink("dummyplot.png") ; unlink("dummyplot_1.png")

}
\seealso{
\code{\link{pdf}}, \code{\link{png}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, March 2017
}
\keyword{file}
