% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDWD.R
\name{readDWD}
\alias{readDWD}
\title{process data from DWD}
\usage{
readDWD(file, format = NA)
}
\arguments{
\item{file}{Name of Zip-File downloaded with \code{\link{dataDWD}},
e.g. "tageswerte_KL_02575_akt.zip".}

\item{format}{Format passed to \code{\link{as.POSIXct}} (see \code{\link{strptime}})
to convert the date/time column to POSIX time format .
If NULL, no conversion is performed (date stays a factor).
If NA, \code{readDWD} tries to find suitable format based on the number of characters.}
}
\value{
data.frame of the desired dataset
}
\description{
Read climate data that was downloaded with \code{\link{dataDWD}}.
}
\note{
These functions are now in the package rdwd and will be removed here!
\url{https://github.com/brry/rdwd}
}
\examples{
# see dataDWD

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jul 2016
}
\seealso{
\code{\link{dataDWD}}
}
\keyword{aplot}

