% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectround.R
\name{rectround}
\alias{rectround}
\title{Rectangles with rounded corners}
\usage{
rectround(xleft, ybottom, xright, ytop, rounding = 0.25, aspcorrect = FALSE,
  bothsame = FALSE, corners = 1:4, npoints = 1000, ...)
}
\arguments{
\item{xleft, ybottom, xright, ytop}{Single numbers with the outer end locations
of the rectangle.}

\item{rounding}{Proportion of the box to round. Recommended to be between
0 and 0.5. DEFAULT: 0.25}

\item{aspcorrect}{Correct for graph aspect ratio? DEFAULT: FALSE}

\item{bothsame}{Set the visual amount of rounding to the same in
both x and y direction? If TRUE,
the proportion relates to the shortest rectangle side.
DEFAULT: FALSE}

\item{corners}{Vector with integers indicating the corners to round.
Starting bottom left, going clockwise. DEFAULT: 1:4}

\item{npoints}{Total number of vertices for the corners. DEFAULT: 1000}

\item{\dots}{Further arguments passed to \code{\link{polygon}},
like col, border, ...}
}
\value{
Final coordinates, invisible
}
\description{
Draw \link{rect}angles with rounded corners via \code{\link{polygon}}
}
\examples{
plot(1:10)  ;  rect(4,2,7,8, border=8)
rectround(4,2,7,8, rounding=0.1)
rectround(4,2,7,8, rounding=0.25) # default
rectround(4,2,7,8, rounding=0.5)
rectround(4,2,7,8, rounding=-0.1, border="red")
rectround(4,2,7,8, rounding=0.8, border="blue")

plot(1:10)  ;  rect(4,2,7,8, border=8)
rectround(4,2,7,8, corners=c(2,4))

plot(1:10, asp=1)  ;  rect(4,2,7,8, border=8)
rectround(4,2,7,8)
rectround(4,2,7,8, aspcorrect=TRUE, border="red") # results depend on asp...

plot(1:10, asp=1.5)  ;  rect(4,2,7,8, border=8)
rectround(4,2,7,8)
rectround(4,2,7,8, aspcorrect=TRUE, border="red") # results depend on asp...

plot(1:10, asp=1)  ;  rect(4,2,7,8, border=8)
rectround(4,2,7,8) # difference only visible if rect is clearly not a square:
rectround(4,2,7,8, bothsame=TRUE, border="red")
rectround(4,2,7,8, bothsame=TRUE, aspcorrect=TRUE, border="blue")

}
\seealso{
\code{\link{textField}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Dec 2017
}
\keyword{aplot}
