% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round0.R
\name{round0}
\alias{round0}
\title{Round numbers with leading and trailing zeros}
\usage{
round0(x, digits = 0, width = digits + ifelse(digits == 0, 2, 3),
  flag = 0, ...)
}
\arguments{
\item{x}{Values}

\item{digits}{Number of digits to keep. DEFAULT: 2}

\item{width}{Total width (number of characters including dot). DEFAULT: digits+2}

\item{flag}{Flag. DEFAULT: "0"}

\item{\dots}{Further arguments passed to \code{\link{formatC}},
except for "format".}
}
\value{
Character string vector
}
\description{
Round numbers and add leading + trailing zeros
}
\examples{
round0( pi*10^(-3:5), 2)
stopifnot(round0(17.3, 2) == "17.30")
round0(7.3)
round0(7.3, 2)

}
\seealso{
\code{\link{formatC}}, \code{\link{sprintf}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jun 2017
}
