% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showPal.R
\name{showPal}
\alias{showPal}
\title{show color palettes}
\usage{
showPal(cex = 4, ...)
}
\arguments{
\item{cex}{Caracter EXpansion size (width of color bar). DEFAULT: 4}

\item{\dots}{Arguments passed to \code{\link{par}}}
}
\description{
Plot examples of the seqential and diverging color palettes in this package.
Do not use \code{rainbow}: \url{https://eagereyes.org/basics/rainbow-color-map}
}
\examples{
showPal()

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Apr 2016
}
\seealso{
\code{\link{seqPal}}, \code{\link{divPal}}, package \code{RColorBrewer}
}
\keyword{color}
\keyword{dplot}

