% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiralDate.R
\name{spiralDate}
\alias{spiralDate}
\title{Spiral graph of time series}
\usage{
spiralDate(dates, values, drange = NULL, months = c("J", "F", "M", "A", "M",
  "J", "J", "A", "S", "O", "N", "D"), add = FALSE,
  zlab = substitute(values), ...)
}
\arguments{
\item{dates}{Dates in ascending order. 
Can be charater strings or \code{\link{strptime}} results, 
as accepted (and coerced) by \code{\link{as.Date}}}

\item{values}{Values to be mapped in color with \code{\link{colPoints}} along seasonal spiral}

\item{drange}{Optional data ragne (analogous to xlim), can be a vector like \code{dates}. DEFAULT: NULL}

\item{months}{Labels for the months}

\item{add}{Add to existing plot? DEFAULT: FALSE}

\item{zlab}{Title of \code{\link{colPointsLegend}}}

\item{\dots}{Further arguments passed to \code{\link{colPoints}}}
}
\value{
invisible data.frame with date, vals, and the plotting coordinates
}
\description{
Plot seasonality of (daily) time serires along spiral
}
\details{
detailsMayBeRemoved
}
\section{Warning}{
 warningMayBeRemoved
}
\examples{

# fake Data
set.seed(42)
time <- as.Date("1985-01-01")+0:5000
vals <- cumsum(rnorm(5001))
vals <- vals + sin(0:5000/366*2*pi)*max(abs(vals))/2

par(mfrow=c(1,2), mar=c(3,3,1,1), mgp=c(2,0.6,0), las=1)
colPoints(time,vals,vals, col=divPal(100), add=FALSE, legend=FALSE,
          lines=TRUE, pch=NA, nint=1, lwd=2)
spiralDate(time,vals, col=divPal(100))
par(mfrow=c(1,1))
spiralDate(time,vals, lines=T, nint=1)

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, May 2016
}
\seealso{
\code{\link{colPoints}}, \code{\link{as.Date}}
}
\keyword{chron}

