% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumatraInitialize.R
\name{sumatraInitialize}
\alias{sumatraInitialize}
\title{Set useful Sumatra PDF Viewer default settings}
\usage{
sumatraInitialize(
  path = "C:/Program Files/RStudio/bin/sumatra",
  openfolder = TRUE
)
}
\arguments{
\item{path}{Folder (not file) that contains "SumatraPDF.exe". 
Hints on automatic path detection would be welcome!
DEFAULT: "C:/Program Files/RStudio/bin/sumatra"}

\item{openfolder}{Logical: Open folder after writing the files?
Uses \code{\link{openFile}()}. DEFAULT: TRUE}
}
\value{
path, invisibly
}
\description{
Set useful Sumatra PDF Viewer default settings. This will likely only work on windows.
At the given \code{path} with "SumatraPDF.exe", this creates two settings files.
Existing files are renamed ("_old_n" appended), not overwritten.\cr
Creates "sumatrapdfrestrict.ini" with \code{SavePreferences = 1} and \code{FullscreenAccess = 1}.\cr
Creates "SumatraPDF-settings.txt" with \code{ShowToc = false} and \code{DefaultDisplayMode = single page}.
\code{UiLanguage} gets filled in by Sumatra itself upon first opening.
}
\examples{
# sumatraInitialize() # only run in interactive mode

}
\seealso{
\code{\link{openPDF}}, \url{https://www.sumatrapdfreader.org/settings/settings.html}, 
\url{https://github.com/sumatrapdfreader/sumatrapdf/blob/master/docs/sumatrapdfrestrict.ini}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, May 2020
}
\keyword{file}
