% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumatraPaths.R
\name{sumatraPaths}
\alias{sumatraPaths}
\title{Get Sumatra PDF Viewer paths}
\usage{
sumatraPaths(open = FALSE)
}
\arguments{
\item{open}{open the folders? DEFAULT: FALSE}
}
\value{
paths
}
\description{
Get Sumatra paths for \code{\link{sumatraInitialize}}. This will only work on windows.
}
\examples{
sumatraPaths()
}
\seealso{
\code{\link{sumatraInitialize}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jan 2024
}
