% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmessage.R
\name{tmessage}
\alias{tmessage}
\alias{twarning}
\alias{tstop}
\title{messages with call trace}
\usage{
tmessage(skip = 0, ...)

twarning(skip = 0, ...)

tstop(skip = 0, ...)
}
\arguments{
\item{skip}{Number of tracing levels to exclude. DEfault: 0}

\item{...}{Passed to \link{message}, \link{warning} or \link{stop}}
}
\value{
NULL, as per \link{message}, \link{warning} or \link{stop}
}
\description{
\link{message}, \link{warning} or \link{stop} with a call trace prepended
}
\examples{
lower <- function(a, s) {tmessage(s, "some stuff with ", a+10); a}
upper <- function(b, skip=0) lower(b+5, skip)
upper(3) # upper -> lower: some stuff with 18
upper(3, skip=1) # no "lower" in trace
upper(3, skip=-1) # upper -> lower -> tmessage: some stuff with 18
}
\seealso{
\link{traceCall} for the generation of the trace
}
