% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toupper1.R
\name{toupper1}
\alias{toupper1}
\title{capitalize words}
\usage{
toupper1(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
character string vector
}
\description{
capitalizes the first letter of character strings using \code{\link{toupper}}
}
\details{
Basically just a one-liner using toupper
}
\examples{
toupper1("berry")
toupper1(c("berRy","likes to code"))

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jul 2016
}
\seealso{
\code{\link{toupper}}, \code{\link{substr}}
}
\keyword{character}

