% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceCall.R
\name{traceCall}
\alias{traceCall}
\title{call stack of a function}
\usage{
traceCall(...)
}
\arguments{
\item{\dots}{Currently ignored}
}
\value{
Character string with the call stack
}
\description{
trace the call stack e.g. for error checking and format output for do.call levels
}
\section{Warning}{
 In \link{do.call} settings with large objects,
                  tracing may take a lot of computing time.
}
\examples{


}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Sep 2016
}
\seealso{
\code{\link{checkFile}} for example usage
}
\keyword{error}
\keyword{programming}

