% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRows.R
\name{addRows}
\alias{addRows}
\title{Add n rows to a data.frame}
\usage{
addRows(df, n, values = NA)
}
\arguments{
\item{df}{Dataframe object}

\item{n}{Number of rows to add}

\item{values}{Values to be used in the new rows. DEFAULT: NA}
}
\value{
A data.frame
}
\description{
simple Helper-Function to add n rows to a data.frame.
}
\examples{

MYDF <- data.frame(A=5:3, B=2:4)
addRows(MYDF, 3)

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jan 2014
}
\seealso{
\code{\link{insertRows}}, \code{\link{data.frame}}, \code{\link{matrix}}, \code{\link{rbind}}
}
\keyword{misc}

