% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anhang.R
\name{anhang}
\alias{anhang}
\title{open the Appendix of Rclick}
\usage{
anhang()
}
\value{
None, opens pdf in default viewer using \code{\link{system2}} or \code{\link{browseURL}}
}
\description{
Open the Appendix of my R handbook found online at
\url{RclickHandbuch.wordpress.com} or directly at
\url{https://dl.dropbox.com/u/4836866/Rclick/Anhang.pdf}.
}
\examples{
# anhang() # excluded from cran check because of external browser opening policy

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jul 2016
}
\seealso{
\code{\link{funSource}}
}

