% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmap.R
\name{bmap}
\alias{bmap}
\title{title}
\usage{
bmap(x = 13.12, y = 52.37, zm = 14, prov = NULL, collapsebg = TRUE, ...)
}
\arguments{
\item{x, y, zm}{passed to \code{leaflet::\link[leaflet]{setView}}}

\item{prov}{named vector of providers. DEFAULT: NULL (nice selection)}

\item{collapsebg}{Collapse background (map) layer selection? DEFAULT: TRUE}

\item{\dots}{Ignored for now}
}
\value{
ReturnValue
}
\description{
description
}
\examples{
if(requireNamespace("leaflet", quietly=TRUE) && 
   requireNamespace("leaflet.extras", quietly=TRUE))
bmap()

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Dec 2022
}
