% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls.R
\name{cls}
\alias{cls}
\title{clear console by function - does not work!}
\usage{
cls()
}
\description{
clear console (CTRL + L) using a function call. Does not work, as \code{rcom} is not available!
}
\examples{

#cls()

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jan 2014
}
\references{
\url{http://r.789695.n4.nabble.com/how-to-clear-screen-in-R-console-td793936.html}
}
\keyword{programming}
\keyword{utilities}

