% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareFiles.R
\name{compareFiles}
\alias{compareFiles}
\title{Compare textfiles for equality}
\usage{
compareFiles(
  file1,
  file2,
  nr = 20,
  startline = 1,
  endline = length(f1),
  quiet = FALSE,
  ...
)
}
\arguments{
\item{file1, file2}{Filenames to be read by \code{\link{readLines}}.}

\item{nr}{number of results printed. DEFAULT: 20}

\item{startline, endline}{start and end lines, e.g. to exclude section that is already compared.}

\item{quiet}{show warnings about file lengths? DEFAULT: FALSE}

\item{\dots}{further arguments passed to \code{\link{readLines}}}
}
\value{
Vector of line numbers that differ, result from \code{\link{head}(..., nr)}
}
\description{
Returns the line numbers where two (text)files differ
}
\examples{

filenames <- system.file(paste0("extdata/versuch",1:2,".txt"), package="berryFunctions")
compareFiles(filenames[1], filenames[2], warn=FALSE)

}
\seealso{
\url{https://text-compare.com/} which I sadly only discovered after writing this function,
         \code{\link{dupes}} for finding duplicate lines, \code{\link{combineFiles}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Aug 2014
}
\keyword{IO}
\keyword{character}
\keyword{file}
