% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFun.R
\name{createFun}
\alias{createFun}
\title{create function framework}
\usage{
createFun(fun, package = "berryFunctions", path = "S:/Dropbox/Public")
}
\arguments{
\item{fun}{Character string or unquoted name. Function that will be crated with identical filename}

\item{package}{Character String with package name. DEFAULT: "berryFunctions"}

\item{path}{Path to package in development (not package name itself) DEFAULT: "S:/Dropbox/Public"}
}
\value{
file name as character string
}
\description{
create a file with a complete (Roxygen) framework for a new function in this package
}
\details{
Tries to open the file in the standard editor for .R files using \code{\link{system2}}
}
\examples{
#createFun("myNewFunction")
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, March 2016
}
\seealso{
\code{\link{system2}}, \code{\link{funSource}}, Roxygen2: 
         \url{https://cran.r-project.org/package=roxygen2/vignettes/rd.html}
}
\keyword{documentation}

