% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFun.R
\name{createFun}
\alias{createFun}
\title{create function framework}
\usage{
createFun(fun, path = ".", open = TRUE)
}
\arguments{
\item{fun}{Character string or unquoted name. Function that will be created with identical filename.}

\item{path}{Path to package in development (including package name itself).
Is passed to \code{\link{packagePath}}. DEFAULT: "."}

\item{open}{Logical: open the file? If several instances of Rstudio are open,
the last one (not necessarily the active one) will be used.
DEFAULT: TRUE}
}
\value{
file name as character string
}
\description{
create a file with a complete (Roxygen) framework for a new function in a package
}
\details{
Tries to open the file in the standard editor for .R files using \code{\link{system2}}
}
\examples{
#createFun("myNewFunction")
}
\seealso{
\code{\link{system2}}, \code{\link{funSource}}, Roxygen2:
         \url{https://cran.r-project.org/package=roxygen2/vignettes/rd.html}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, March 2016
}
\keyword{documentation}
