% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPres.R
\name{createPres}
\alias{createPres}
\title{Create .Rnw presentation template}
\usage{
createPres(
  presname = "pres",
  dir = "presentation",
  path = ".",
  asp = 169,
  navbullets = FALSE,
  bgblack = FALSE,
  open = TRUE
)
}
\arguments{
\item{presname}{Name of .Rnw file to be created. DEFAULT: "pres"}

\item{dir}{Name of directory that will contain .Rnw file and
fig_extern folder. "_1" will be appended if already existing,
see \code{\link{newFilename}}. DEFAULT: "presentation"}

\item{path}{Location of \code{dir}. Passed to \code{\link{setwd}}.
DEFAULT: "."}

\item{asp}{Number to set as aspectratio. 43 for old 4:3 format. 
Possible values: 169, 1610, 149, 54, 43, 32. 
\bold{note}: if you set this, remember to change the default \code{fig.width}.
DEFAULT: 169 (16:9 format)}

\item{navbullets}{Logical: include navigation slide bullet points in header?
This only takes effect when there are subsections.
DEFAULT: FALSE}

\item{bgblack}{Logical: set a black background instead of a white one?
Requires all R graphics fg and bg colors to be changed!
See "How to avoid death By PowerPoint" at 11:49 minutes 
\url{https://youtu.be/Iwpi1Lm6dFo?t=11m49s}.
Change colors manually in the Rnw files searching for 
\code{bg=}, \code{linkcolor=}, \code{urlcolor=}
in the preamble and \code{color} right after \code{begin document}.
DEFAULT bgblack: FALSE}

\item{open}{Logical: run \code{\link{openFile}}? DEFAULT: TRUE}
}
\description{
Create folder with .Rnw presentation template and fig_extern folder.
}
\examples{
\dontrun{
createPres("Berry_Conference")
}

}
\seealso{
\code{\link{createFun}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Mar 2017
}
\keyword{file}
