% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berryFunctions-package.R
\name{dataDWD}
\alias{dataDWD}
\alias{readDWD}
\title{Get climate data from the German Weather Service (DWD) FTP-server.}
\description{
\code{dataDWD} and \code{readDWD} have been greatly improved and are now in the package \code{rdwd}.
This provides a much better infrastructure for file selection and downloading. 
See \url{https://github.com/brry/rdwd#rdwd} for usage instructions.
}

