% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataStr.R
\name{dataStr}
\alias{dataStr}
\title{str of datasets}
\usage{
dataStr(package = "datasets", ...)
}
\arguments{
\item{package}{package. DEFAULT: "datasets"}

\item{\dots}{other arguments passed to \code{\link{data}}}
}
\value{
NULL. prints via \code{\link{message}} in a for loop.
}
\description{
Print the \code{\link{str}} of each dataset returned by \code{\link{data}},
by default in the package \code{\link{datasets}}
}
\examples{

# dataStr()

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, November 2015, in search of good datasets for teaching
}
\seealso{
\code{\link{str}}
}
\keyword{documentation}
\keyword{print}

