% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileCheck.R
\name{fileCheck}
\alias{fileCheck}
\title{check file existance}
\usage{
fileCheck(file, fun = stop)
}
\arguments{
\item{file}{Filename as character string to be checked for existence.}

\item{fun}{One of the functions \code{\link{stop}}, \code{\link{warning}}, or \code{\link{message}}. DEFAULT: stop}
}
\value{
TRUE/FALSE, invisibly
}
\description{
check whether a file exists and give a useful error/warning/message
}
\examples{
is.error( fileCheck("FileThatDoesntExist.txt") )
fileCheck("FileThatDoesntExist.txt", fun=warning)
fileCheck("FileThatDoesntExist.txt", fun=message)
is.error( fileCheck("FileThatDoesntExist.txt", fun=MyWarn) ) # nonexisting function

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, May 2016
}
\seealso{
\code{\link{file.exists}}
}
\keyword{file}

