% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funSource.R
\name{funSource}
\alias{funSource}
\title{Source code of a function}
\usage{
funSource(x, character.only = FALSE, trydirect = TRUE)
}
\arguments{
\item{x}{function name, with or without quotation marks}

\item{character.only}{If TRUE, look for SomeFun instead of MyFun if
MyFun <- "SomeFun". DEFAULT: FALSE}

\item{trydirect}{If TRUE, try direct url to file \code{x.R}. DEFAULT: TRUE}
}
\value{
link that is also opened
}
\description{
open source code of a function in a loaded package on github.com/cran or github.com/wch/r-source
}
\note{
This is not finished yet...
}
\examples{
library("berryFunctions")
funSource(colPoints) 
funSource("head") 
funSource("require", trydirect=FALSE) 

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jan 2016
}

