% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funSource.R
\name{funSource}
\alias{funSource}
\title{Source code of a function}
\usage{
funSource(x, character.only = FALSE)
}
\arguments{
\item{x}{function name, with or without quotation marks}

\item{character.only}{If TRUE, look for SomeFun instead of MyFun if  MyFun <- "SomeFun". DEFAULT: FALSE}
}
\value{
link that is also opened
}
\description{
open github.com/cran source code of a function in a package
}
\note{
This is not finished yet...
}
\examples{
# ToDo: add message: "function also found in: " 

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jan 2016
}

