% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funTinn.R
\name{funTinn}
\alias{funTinn}
\title{Open function in TinnR}
\usage{
funTinn(name)
}
\arguments{
\item{name}{Name of function or object to be opened with the program associated with .r-files. In my case, the editor Tinn-R}
}
\description{
Opens function or object in external editor with an R command
}
\examples{

\dontrun{
## Rcmd check --as-cran doesn't allow opening external devices,
## so this example is excluded from running in the checks.
funTinn(boxplot.default)
}

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Aug 2014
}
\seealso{
\code{\link{edit}}, \url{http://stackoverflow.com/questions/13873528}
}
\keyword{connection}

