% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColumn.R
\name{getColumn}
\alias{getColumn}
\title{get column from data.frame}
\usage{
getColumn(x, df, trace = TRUE)
}
\arguments{
\item{x}{Column name to be subsetted}

\item{df}{dataframe object}

\item{trace}{Logical: Add function call stack to the message? DEFAULT: TRUE
WARNING: in do.call settings with large objects,
tracing may take a lot of computing time.}
}
\value{
Vector with values in the specified column
}
\description{
Extract columns if they are given in a data frame
}
\examples{
getColumn(Air.Flow, stackloss)
is.error(getColumn(Acid, stackloss))

upper <- function(x) getColumn(x, stackloss)
upper(Water.Temp)
# upper(Water) 

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Sep 2016
}
\seealso{
\code{\link{subset}}
}

