% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsMem.R
\name{lsMem}
\alias{lsMem}
\title{Show memory size of objects in MB}
\usage{
lsMem(n = 6, pos = 1, ...)
}
\arguments{
\item{n}{Number of Objects to be shown separately. The rest is combined into "sum rest". DEFAULT: 6}

\item{pos}{Environment where \code{\link{ls}} looks for objects.}

\item{\dots}{Further arguments passed to \code{\link{ls}}}
}
\value{
Named vector with object sizes in MB (MegaBytes)
}
\description{
Show memory size of the biggest objects in MB. Helps you find the biggest memory killers.
}
\examples{

\dontrun{
## excluded from CRAN check  - I forgot why, but there's probably a good reason
lsMem()
}

}
\references{
\url{https://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session}
}
\seealso{
\code{\link{object.size}}, \code{\link{ls}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Feb 2014
}
\keyword{file}
\keyword{programming}
