% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastec.R
\name{pastec}
\alias{pastec}
\title{Paste with collapse = ", "}
\usage{
pastec(..., sep = " ", collapse = ", ")
}
\arguments{
\item{sep}{Character string to separate single strings. DEFAULT: " "}

\item{collapse}{Character string between combined strings. DEFAULT: ", "}

\item{\dots}{Object(s) to be \code{\link{paste}d} to a character vector}
}
\value{
Single character string
}
\description{
Helper function \code{\link{paste}} with collapse = ", "
}
\examples{

listoferrors <- c("filetype", "header", "nonemptyline")
message("The following entities were corrupted:\\n", pastec(listoferrors))
pastec("Part1", c("Part2", "Part3"), letters[1:3])

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, April 2015
}
\seealso{
\code{\link{paste}}, \code{\link{paste0}}
}
\keyword{character}

