% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require2.R
\name{require2}
\alias{library2}
\alias{require2}
\title{install.package and require}
\usage{
require2(name, ...)
}
\arguments{
\item{name}{Name of the package(s). Can be qouted, must not.}

\item{\dots}{Arguments passed to \code{\link{install.packages}} like \code{lib}, \code{repos} etc.}
}
\value{
Cats help instruction.
}
\description{
install and load a package. If a package is not available, it is installed before being loaded
}
\note{
Passing a vector with packages will work, but give some warnings.
}
\examples{

\dontrun{
## Excluded fom CRAN checks. Package installation on server is unnecessary.
require2(ada)
}

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, 2014
}
\seealso{
\code{\link{install.packages}}, \code{\link{require}}
}
\keyword{package}

