% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round0.R
\name{round0}
\alias{round0}
\title{Round numbers with leading and trailing zeros}
\usage{
round0(
  x,
  digits = 0,
  pre = 2,
  width = digits + pre + ifelse(digits == 0, 0, 1),
  flag = 0,
  ...
)
}
\arguments{
\item{x}{Value(s)}

\item{digits}{Number of digits (after decimal separator) to keep. DEFAULT: 0}

\item{pre}{Minimum number of characters before the decimal separator. DEFAULT: 2}

\item{width}{Total width (number of characters including dot).
DEFAULT: digits+pre (+1 if needed)}

\item{flag}{Flag. Could be "" for spaces. DEFAULT: "0"}

\item{\dots}{Further arguments passed to \code{\link{formatC}},
except for "format".}
}
\value{
Character string vector
}
\description{
Round numbers and add leading + trailing zeros
}
\examples{
round0( pi*10^(-3:5), 2)
stopifnot(round0(17.3, 2) == "17.30")
round0(7.3)
round0(c(7.3,777.1234), 2)
round0(c(0.2,7.3,12.8), 2, pre=1)
round0(c(0.2,7.3,12.8), 1, pre=3, flag="") # spaces instead of zeros

}
\seealso{
\code{\link{formatC}}, \code{\link{sprintf}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jun 2017
}
