% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqPal.R
\name{seqPal}
\alias{seqPal}
\title{Sequential color palette}
\usage{
seqPal(n = 12, reverse = FALSE, alpha = 1, extr = FALSE, yb = FALSE,
  yr = FALSE, colors = NULL, ...)
}
\arguments{
\item{n}{Number of colors. DEFAULT: 12}

\item{reverse}{Reverse colors? DEFAULT: FALSE}

\item{alpha}{Transparency (0=transparent, 1=fully colored). DEFAULT: 1}

\item{extr}{Should colors span possible range more extremely? 
If TRUE, it has very light yellow and very dark blue values included, 
using the result from \code{RColorBrewer::brewer.pal(9, "YlGnBu")}. DEFAULT: FALSE}

\item{yb}{Should colors be in yellow-blue instead of the internal (nice) default? DEFAULT: FALSE}

\item{yr}{Should colors be in yellow-red instead of the default? DEFAULT: FALSE}

\item{colors}{If not NULL, a color vector used in \code{\link{colorRampPalette}}. DEFAULT: NULL}

\item{\dots}{Further arguments passed to \code{\link{colorRamp}}}
}
\value{
Character string vector with color names
}
\description{
Sequential color palette from yelow to red or yellow to blue or custom colors.
}
\examples{

op <- par(mfrow=c(11,1), mar=c(0,0,0,0))
for(n in c(3,7,11,12,300)) plot(rep(1,n), pch=16, cex=5, col=seqPal(n), xaxt="n")
text(150, 1, "berryFunctions::seqPal default palette")
plot(rep(1,300), pch=16, cex=5, col=seqPal(300, extr=TRUE), xaxt="n")
text(150, 1, "extr=TRUE")
plot(rep(1,12),  pch=16, cex=5, col=seqPal(alpha=0.4), xaxt="n")
text(6, 1, "semi-transparency with alpha=0.4")
plot(rep(1,12),  pch=16, cex=5, col=seqPal(rev=TRUE), xaxt="n")
plot(rep(1,300), pch=16, cex=5, col=seqPal(300,yb=TRUE), xaxt="n")
plot(rep(1,300), pch=16, cex=5, col=seqPal(300,yr=TRUE), xaxt="n")
plot(rep(1,300), pch=16, cex=5, col=seqPal(300,col=c("darkblue","green","orange")), xaxt="n")
par(op)

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jan 2016
}
\seealso{
\code{\link{divPal}}, \code{\link{addAlpha}}, \code{\link{colorRampPalette}}, package RColorBrewer
}
\keyword{color}
\keyword{dplot}

