% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.R
\name{showPal}
\alias{showPal}
\title{show color palettes}
\usage{
showPal(cex = 4, ...)
}
\arguments{
\item{cex}{Character EXpansion size (width of color bar). DEFAULT: 4}

\item{\dots}{Arguments passed to \code{\link{par}}}
}
\description{
Plot examples of the sequential and diverging color palettes in this package.
Do not use \code{rainbow}: \url{https://eagereyes.org/basics/rainbow-color-map}
}
\examples{
showPal()
}
\seealso{
\code{\link{seqPal}}, \code{\link{divPal}}, \code{\link{catPal}}, 
package \code{RColorBrewer}, 
\url{https://blog.datawrapper.de/which-color-scale-to-use-in-data-vis/}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Apr 2016
}
\keyword{color}
\keyword{dplot}
