% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortDF.R
\name{sortDF}
\alias{sortDF}
\title{sort dataframes by column}
\usage{
sortDF(df, col, decreasing = TRUE, ...)
}
\arguments{
\item{df}{Data.frame to be sorted}

\item{col}{Column (index or (un)qouted name) to be sorted by}

\item{decreasing}{Logical: should highest value be on top? 
DEFAULT: TRUE (unlike \code{\link{order}}!)}

\item{\dots}{Further arguments passed to \code{\link{order}}, like eg \code{na.last or method}}
}
\value{
data.frame
}
\description{
sort a data.frame by column - basically just a wrapper for order
}
\examples{

sortDF(USArrests[USArrests$Murder>11,], "Assault")
sortDF(USArrests[USArrests$Murder>11,], 3)

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, June 2015
}
\seealso{
\code{\link{sort}}, \code{\link{order}}
}
\keyword{arith}
\keyword{manip}
\keyword{univar}

