% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{timer}
\alias{timer}
\title{Timer alarm}
\usage{
timer(interval = 20, n = 15, write = FALSE)
}
\arguments{
\item{interval}{\code{\link{alarm}} interval in seconds. DEFAULT: 20}

\item{n}{number of alarm signals to be given. DEFAULT: 15}

\item{write}{Should the actual estimated time be written for overhead computing time control purposes? DEFAULT: FALSE}
}
\value{
none
}
\description{
Beeps in a given interval and gives a progress bar in the console
}
\details{
defaults to practice useR lightning talks: 15 slides, each shown 20 secs, change automatically
}
\examples{

\dontrun{## Skip time consuming checks on CRAN
timer(interval=0.5, n=3)
timer(interval=0.2, n=8, write=TRUE) # a slight deviation occurs for a large n
# timer() # to practice lightning talks at useR! conferences
}

}
\references{
\url{http://user2015.math.aau.dk/lightning_talks}
}
\seealso{
\code{\link{alarm}}, \code{\link{Sys.sleep}}, \code{\link{txtProgressBar}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, June 2015
}
\keyword{chron}
\keyword{utilities}
