% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryStack.R
\name{tryStack}
\alias{tryStack}
\title{try an expression, returning the error stack}
\usage{
tryStack(expr, silent = FALSE, tracewarnings = TRUE, file = "",
  removetry = TRUE)
}
\arguments{
\item{expr}{Expresssion to try, potentially wrapped in curly braces if 
spanning several commands.}

\item{silent}{Logical: Should error message + stack printing be suppressed?
DEFAULT: FALSE}

\item{tracewarnings}{Logical: Should warnings be traced as well?
They will still be printed as regular warnings, 
but with trace stack. DEFAULT: TRUE}

\item{file}{File name passed to \code{\link{cat}}. 
If given, Errors will be appended to the file after two empty lines. 
if tracewarnings=TRUE and file!="", warnings will not be shown, 
but also appended to the file.
This is useful in lapply simulation runs.
DEFAULT: "" (catted to the console)}

\item{removetry}{Logical: should all stack entries matching typical tryCatch
expressions be removed? Unless the call contains customized
\code{\link{tryCatch}} code, this can be left to the DEFAULT: TRUE}
}
\value{
Value of \code{expr} if evaluated successfully. If not, an invisible 
object of class "try-error" as in \code{\link{try}} with the stack in \code{object[2]}.
}
\description{
As in \code{\link{try}}, the result of an expression if it works.
If it fails, no error is thrown, but an invisible try-error class object is 
returned and a message \code{\link{cat}ted} to the console. Suppress the latter with silent=TRUE\cr
Unlike \code{\link{try}}, \code{tryStack} also returns the calling stack to ease debugging.
}
\examples{

lower <- function(a) a+10
middle <- function(b) {plot(b, main=b) ; warning("fake warning, b = ", b); lower(b) }
upper <- function(c) {cat("printing c:", c, "\\n") ; middle(c)}
d <- upper(42)
d
rm(d)

\dontrun{ ## intentional error
d <- upper("42")                # error, no d creation 
traceback()                     # calling stack, but only in interactive mode
}


d <- try(upper("42"), silent=TRUE)      # d created
cat(d)                                  # has error message, but no traceback
inherits(d, "try-error")                # use for coding


d <- tryStack(upper("42"), silent=TRUE) # like try, but with traceback, even for warnings
cat(d)
cat(tryStack(upper("42"), silent=TRUE)) # level 1 different, but correct

d <- tryStack(upper("42"), silent=TRUE, tracewarnings=FALSE) # don't touch warnings

tryStack(upper(42)) # returns normal output, but warnings are easier to debug

stopifnot(inherits(d, "try-error"))
stopifnot(tryStack(upper(42))==52)


myfun <- function(k) tryStack(upper(k), silent=TRUE)
d <- myfun(42)
cat(d)
d <- myfun("42")
cat(d)
 
myfun <- function(k) tryStack(stop("oh oh"))
d <- myfun(42)

d <- tryStack(myfun(4) ) # nested calls get weird
cat(d)                   # They make that empty in the second run, for example.

myfun <- function(k) cat( tryStack(upper("42")) )
myfun(42) 
d <- tryStack(myfun(42) ) 
rm(myfun)


myfun1 <- function(k) try(upper(k), silent=TRUE)
d <- myfun1(42)
d <- myfun1("42") ; cat(d) # regular try output
myfun2 <- function(k) tryStack(myfun1(k), tracewarnings=FALSE, silent=TRUE)
d <- myfun2(42)
d <- myfun2("42")
cat(d) # empty stack - because of try, no real error happened

myfun1 <- function(k) tryStack(upper(k), silent=TRUE)
d <- myfun2("42")
cat(d)            # same result again! (ToDo: check if d[2] is the culprit)
 

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Nov 2016
}
\seealso{
\code{\link{try}}, \code{\link{traceCall}},
\url{http://r.789695.n4.nabble.com/Stack-trace-td4021537.html},
\url{http://stackoverflow.com/questions/15282471/get-stack-trace-on-trycatched-error-in-r},
\url{http://stackoverflow.com/questions/1975110/printing-stack-trace-and-continuing-after-error-occurs-in-r},
\url{http://stackoverflow.com/questions/16879821/save-traceback-on-error-using-trycatch}
}
\keyword{error}
\keyword{programming}

