% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearSample.R
\name{yearSample}
\alias{yearSample}
\title{Nonrandom year with sample}
\usage{
yearSample(year)
}
\arguments{
\item{year}{4 digit numerical year.}
}
\value{
\code{\link{cat}}s command into the console that can be copypasted to anyone's R script.
}
\description{
Nerdy way to wish someone a happy new year by using sample
}
\details{
Nerdy way to wish someone a happy new year, eg:\cr
Have a great \cr
\code{set.seed(1244); sample(0:9,4,T)}
}
\examples{

yearSample(2016)
# Have a nerdy
set.seed(12353); sample(0:9, 4, replace=TRUE)

}
\seealso{
\code{\link{nameSample}} to impress with "randomly" finding a name,
         \code{\link{set.seed}}, \code{\link{sample}}, \code{\link{letters}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, April 2014
}
